backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.9.0
pydantic-settings==2.6.0
pillow==11.0.0
